+++ spine-ios/Example/Spine iOS Example/DebugRendering.swift
@@ -17,8 +17,7 @@ struct DebugRendering: View {
         ZStack {
             Color.red.ignoresSafeArea()
             SpineView(
-                atlasFile: "spineboy.atlas",
-                skeletonFile: "spineboy-pro.skel",
+                from: .bundle(atlasFileName: "spineboy.atlas", skeletonFileName: "spineboy-pro.skel"),
                 controller: model.controller,
                 mode: .fit,
                 alignment: .center