+++ spine-ios/Example/Spine iOS Example/IKFollowing.swift
@@ -15,8 +15,7 @@ struct IKFollowing: View {
         
     var body: some View {
         SpineView(
-            atlasFile: "spineboy.atlas",
-            skeletonFile: "spineboy-pro.skel",
+            from: .bundle(atlasFileName: "spineboy.atlas", skeletonFileName: "spineboy-pro.skel"),
             controller: model.controller,
             alignment: .centerLeft
         )