+++ spine-ios/Example/Spine iOS Example/SimpleAnimation.swift
@@ -23,8 +23,11 @@ struct SimpleAnimation: View {
     
     var body: some View {
         SpineView(
-            atlasFile: "spineboy.atlas",
-            skeletonFile: "spineboy-pro.skel",
+            from: .bundle(atlasFileName: "spineboy.atlas", skeletonFileName: "spineboy-pro.skel"),
+//            from: .http(
+//                atlasURL: URL(string: "https://github.com/denrase/spine-runtimes/raw/spine-ios/spine-ios/Example/Spine%20iOS%20Example/Assets/spineboy/spineboy.atlas")!,
+//                skeletonURL:  URL(string: "https://github.com/denrase/spine-runtimes/raw/spine-ios/spine-ios/Example/Spine%20iOS%20Example/Assets/spineboy/spineboy-pro.skel")!
+//            ),
             controller: controller,
             mode: .fit,
             alignment: .center